requirements.txt
google-genai==1.18.0
langchain==0.3.25
langchain-google-genai==2.1.5
python-dotenv==1.1.0